requirements.txt
gradio>=5.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-astradb>=0.6.0
langchain-chroma>=0.1.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
pymupdf>=1.24.0
pypdf>=4.3.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.1
